+++ xunsearch/app/util/mongo_query.php
@@ -26,7 +26,8 @@ $options = [
 
 $query = new MongoDB\Driver\Query($filter, $options);
 
-$manager = new MongoDB\Driver\Manager("mongodb://root:forester@forester-mongo:27017");
+$uri = getenv("MONGODB_URI");
+$manager = new MongoDB\Driver\Manager($uri);
 $readPreference = new MongoDB\Driver\ReadPreference(MongoDB\Driver\ReadPreference::RP_PRIMARY);
 $cursor = $manager->executeQuery('cmccb2b.BidNotice', $query, $readPreference);