+++ Changelog.md
@@ -8,6 +8,7 @@
 ## TODO
 
 - Flask要简化uwsgi，改为gunicron方式
+- cronjobs镜像改造为jobservices模式
 
 ---